requirements.txt
apache-airflow[postgres, amazon] == 2.3.3
fastparquet==0.8.1
